FEATURE: add look-ahead too main camera delete hardcoded lints in favor of cargo-cranky remove some ahp:: glob imports add new AspenCollisionLayer struct with preset layers     - reasoning about this has been very hard for me     - this new set with some documentation should be better hopefully undo moving clippy config out of workspace root     - bevy has some weird interactions with clippys type complexity lint use f32s in console commands, no more casting types remove #[allow(type_complexity)] in favor of clippy config remove background tiling material, breaks wasm and doenst work anyways refactor touchpad update systems custom game difficulty setting, can configure the bits proper error handling in the save_settings function

+++ src/console/commands.rs
@@ -8,9 +8,9 @@ pub struct SpawnWeaponCommand {
     /// type of w to spawn
     pub weapon_type: String,
     /// z transform of weapon
-    pub loc_x: Option<i64>,
+    pub loc_x: Option<f32>,
     /// y transform
-    pub loc_y: Option<i64>,
+    pub loc_y: Option<f32>,
     /// Number of times to spawn
     pub amount: Option<i32>,
     /// spawn at/near player
@@ -25,9 +25,9 @@ pub struct SpawnEnemyCommand {
     /// type of thing to spawn
     pub enemy_type: String,
     /// x transform
-    pub loc_x: Option<i64>,
+    pub loc_x: Option<f32>,
     /// y transform
-    pub loc_y: Option<i64>,
+    pub loc_y: Option<f32>,
     /// Number of times to spawn
     pub amount: Option<i32>,
     /// spawn at/near player
@@ -39,7 +39,7 @@ pub struct SpawnEnemyCommand {
 #[command(name = "teleport")]
 pub struct TeleportPlayerCommand {
     /// x pos to teleport too
-    pub loc_x: i64,
+    pub loc_x: f32,
     /// y pos to teleport too
-    pub loc_y: i64,
+    pub loc_y: f32,
 }
